refactor list iocs
